Changed Code to No GUI Class
Had to scratch the whole MyGUI Class as I was having trouble accessing and updating the loading bar widget outside of the class. I needed the download button function to take the download_video function and use the pytube output_path function. Which gives me the information on the total size of the file and the remaining bytes of transfer and such. Which I use to increment the loading bar with those values. Other than that I got a progress bar to show and work.